Bug fixed and created SignUp page
